refactor: use double incident distance from updated library
RouteIncident.distanceFromStartMetres is now Double in the external nav
API library, so the explicit .toDouble() cast in ExtNavRouteIncidentMapper
is no longer needed. Update test helper argument type and all call-site
literals to Double accordingly.

+++ core/navigation/src/androidMain/kotlin/me/matsumo/onenavi/core/navigation/extnav/ExtNavRouteIncidentMapper.kt
@@ -60,7 +60,7 @@ internal object ExtNavRouteIncidentMapper {
         distanceOffsetMeters: Double,
         totalGeometryMetres: Double,
     ): RouteIncidentMarker {
-        val geometryDistanceFromStartMeters = (distanceFromStartMetres.toDouble() + distanceOffsetMeters)
+        val geometryDistanceFromStartMeters = (distanceFromStartMetres + distanceOffsetMeters)
             .coerceIn(0.0, totalGeometryMetres)
 
         return RouteIncidentMarker(
